fix build for reconverse
