+++ src/cmi-shmem-common.h
@@ -1,6 +1,7 @@
 #ifndef CMI_SHMEM_COMMON_HH
 #define CMI_SHMEM_COMMON_HH
 
+#include "converse_internal.h"
 #include "cmishmem.h"
 
 #include <array>
